advertiser column

+++ backend/server/src/schema.rs
@@ -2,15 +2,15 @@ infer_schema!("dotenv:DATABASE_URL");
 
 // Dummy tables for derived queries
 table! {
-    use diesel::types::{Integer, Text};
+    use diesel::types::{BigInt, Text};
     aggregate_targetings (count) {
-        count -> Integer,
+        count -> BigInt,
         targeting_type -> Text,
     }
 }
 
 #[derive(QueryableByName, Debug, Clone)]
 pub struct AggregateTargeting {
-    pub count: i32,
+    pub count: i64,
     pub targeting_type: String,
 }
